Add `ZabbixApiInterface`
